Changes to reduce mem copies, improve speedup
Moved the #pragma omp task call after if statements that check if the adjacent spot is a valid spot and is empty, to get rid of thread launches that aren't going to actually make it to the recursive call. Then, found a way to not copy local_arr (the array holding the contents of the board) every task call. These changes improved speedup.